fix: widen synapse from_index to u32 for neat-core 0.1.46
neat-core Issue #177 narrowed SynapseData::from_index from u32 to u16.
rust_scorer's GPU mirror SynapseGpu (and its WGSL shaders) keep from_index
as u32, so the direct assignment in forward_mse_batched.rs no longer type
checks (E0308: expected u32, found u16). Widen with u32::from, matching the
existing conversions for squash_type and num_synapses on adjacent lines.

This unblocks the "Unit tests + coverage" check on
stSoftwareAU/NEAT-AI-Examples#606, which builds rust_scorer against the
latest neat-core Develop.

Vibe-Coder-Run-Id: vibe-mqqdj58s-b50a27

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,9 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core Issue #177 narrowed SynapseData::from_index to u16;
+                // the GPU mirror (and its WGSL shaders) keep it as u32, so widen.
+                from_index: u32::from(s.from_index),
             });
         }
